+++ R/read_fwf.R
@@ -37,7 +37,8 @@ read_fwf <- function(file, col_positions = fwf_empty(file, skip, n = guess_max),
                      locale = default_locale(), na = c("", "NA"),
                      comment = "", trim_ws = TRUE, skip = 0, n_max = Inf,
                      guess_max = min(n_max, 1000), progress = show_progress(),
-                      show_col_types = should_show_types(),
+                     num_threads = readr_threads(),
+                     show_col_types = should_show_types(),
                      lazy = TRUE, skip_empty_rows = TRUE) {
   if (edition_first()) {
     ds <- datasource(file, skip = skip, skip_empty_rows = skip_empty_rows)
@@ -78,7 +79,7 @@ read_fwf <- function(file, col_positions = fwf_empty(file, skip, n = guess_max),
   vroom::vroom_fwf(file, col_positions = col_positions, col_types = col_types,
     locale = locale, na = na, comment = comment, trim_ws = trim_ws, skip = skip,
     n_max = n_max, guess_max = guess_max, show_col_types = show_col_types,
-    progress = progress, altrep = lazy)
+    progress = progress, altrep = lazy, num_threads = num_threads)
 }
 
 #' @rdname read_fwf